+++ src/main/resources/templates/employee/edit.html
@@ -99,7 +99,7 @@
     </style>
 </head>
 <body>
-<form th:action="@{/employee/edit}" th:object="${employee}" method="post">
+<form th:action="@{/employees/edit}" th:object="${employee}" method="post">
     <h2>Chỉnh sửa thông tin cá nhân</h2>
 
     <div class="form-grid">
@@ -153,7 +153,7 @@
 
     <div class="button-group">
         <button type="submit" class="btn-submit">Cập nhật</button>
-        <a th:href="@{/employee/profile}" class="btn-back">Trở về</a>
+        <a th:href="@{/employees/profile}" class="btn-back">Trở về</a>
     </div>
 </form>
