+++ src/main/resources/templates/employee/profile.html
@@ -134,7 +134,7 @@
 
     <div class="actions">
         <button onclick="window.history.back()">Quay về</button>
-        <button onclick="window.location.href='/employee/edit'">Cập nhật thông tin</button>
+        <button onclick="window.location.href='/employees/edit'">Cập nhật thông tin</button>
         <button onclick="window.location.href='/auth/change-password'">Thay đổi mật khẩu</button>
     </div>
 </div>
